refactor:规范了字段数据类型

+++ 01_Core/Models/ModbusRequestParams.cs
@@ -10,7 +10,7 @@ namespace _01_Core.Models
 {
     public class ModbusRequestParams
     {
-        public int SlaveId { get; set; } = 1;
+        public byte SlaveId { get; set; } = 1;
         public FunctionCode FunctionCode { get; set; }=FunctionCode.ReadHoldingRegisters;
         public int StartAddress { get; set; } = 0;
         public int Count { get; set; } = 10;
